Extract multi-selection menu to separate view
