+++ Movie DB/Views/Library/LibraryToolbar.swift
@@ -19,27 +19,12 @@ struct LibraryToolbar: ToolbarContent {
     @Binding var config: LibraryViewModel
     var editMode: Binding<EditMode>?
     @Binding var selectedMediaObjects: Set<Media>
-    @State private var isShowingDeleteAlert: Bool = false
     
     let filterImageReset = "line.horizontal.3.decrease.circle"
     let filterImageSet = "line.horizontal.3.decrease.circle.fill"
     
     var filterImageName: String {
         filterSetting.isReset ? filterImageReset : filterImageSet
-    }
-    
-    private var isEditing: Bool { editMode?.wrappedValue.isEditing ?? false }
-    
-    private var areAllFavorite: Bool {
-        !selectedMediaObjects.isEmpty && selectedMediaObjects.allSatisfy(\.isFavorite)
-    }
-    
-    private var areAllOnWatchlist: Bool {
-        !selectedMediaObjects.isEmpty && selectedMediaObjects.allSatisfy(\.isOnWatchlist)
-    }
-    
-    private var areAllSelected: Bool {
-        !selectedMediaObjects.isEmpty && selectedMediaObjects.count == MediaLibrary.shared.mediaCount()
     }
     
     var body: some ToolbarContent {
@@ -52,21 +37,8 @@ struct LibraryToolbar: ToolbarContent {
     private var moreMenu: some ToolbarContent {
         ToolbarItem(placement: .navigationBarLeading) {
             Menu {
-                Button {
-                    withAnimation {
-                        editMode?.wrappedValue = isEditing ? .inactive : .active
-                    }
-                } label: {
-                    Label(
-                        isEditing ? Strings.Generic.editButtonLabelDone : Strings.Library.menuSelectLabel,
-                        systemImage: "checkmark.circle"
-                    )
-                }
-                if isEditing {
-                    Section {
-                        multiSelectActions
-                    }
-                }
+                MultiSelectionMenu(selectedMediaObjects: $selectedMediaObjects)
+                    .environment(\.editMode, editMode)
                 Section {
                     Button {
                         config.activeSheet = .filter
@@ -87,10 +59,12 @@ struct LibraryToolbar: ToolbarContent {
     
     @ToolbarContentBuilder
     private var multiSelectDoneButton: some ToolbarContent {
-        if isEditing {
+        if editMode?.wrappedValue.isEditing == true {
             ToolbarItem(placement: .topBarTrailing) {
                 Button {
-                    dismissEditing()
+                    withAnimation {
+                        editMode?.wrappedValue = .inactive
+                    }
                 } label: {
                     Text(Strings.Generic.editButtonLabelDone)
                         .bold()
@@ -108,123 +82,6 @@ struct LibraryToolbar: ToolbarContent {
                 Image(systemName: "plus")
             }
             .accessibilityIdentifier("add-media")
-        }
-    }
-}
-
-// MARK: - Multi-Selection Actions
-private extension LibraryToolbar {
-    @ViewBuilder
-    var multiSelectActions: some View {
-        selectAllButton
-        Group {
-            AddMultipleToListMenu(mediaObjects: selectedMediaObjects) {
-                dismissEditing()
-            }
-            addToWatchlistButton
-            addToFavoritesButton
-            reloadButton
-            deleteButton
-            // TODO: Multi-select for lists details (remove)
-        }
-        .disabled(selectedMediaObjects.isEmpty)
-    }
-    
-    var selectAllButton: some View {
-        Button {
-            do {
-                if areAllSelected {
-                    selectedMediaObjects = []
-                } else {
-                    selectedMediaObjects = Set(try managedObjectContext.fetch(Media.fetchRequest()))
-                }
-            } catch {
-                Logger.coreData.error("Failed to fetch all media objects: \(error)")
-            }
-        } label: {
-            if areAllSelected {
-                Text(Strings.Library.multiDeselectAll)
-            } else {
-                Text(Strings.Library.multiSelectAll)
-            }
-        }
-    }
-    
-    var addToWatchlistButton: some View {
-        Button {
-            let isOnWatchlist = !areAllOnWatchlist
-            for media in selectedMediaObjects {
-                media.isOnWatchlist = isOnWatchlist
-            }
-            dismissEditing()
-        } label: {
-            if areAllOnWatchlist {
-                Label(Strings.Detail.menuButtonRemoveFromWatchlist, systemImage: "bookmark.slash.fill")
-            } else {
-                Label(Strings.Detail.menuButtonAddToWatchlist, systemImage: "bookmark.fill")
-            }
-        }
-    }
-    
-    var addToFavoritesButton: some View {
-        Button {
-            // If there is at least one media which is not favorited yet, favorite all medias
-            let isFavorite = !areAllFavorite
-            for media in selectedMediaObjects {
-                media.isFavorite = isFavorite
-            }
-            dismissEditing()
-        } label: {
-            // Favorite is the default action if the favorite statuses are mixed
-            if areAllFavorite {
-                Label(Strings.Detail.menuButtonUnfavorite, systemImage: "heart.fill")
-            } else {
-                Label(Strings.Detail.menuButtonFavorite, systemImage: "heart")
-            }
-        }
-    }
-    
-    var reloadButton: some View {
-        Button {
-            Task {
-                for media in selectedMediaObjects {
-                    do {
-                        try await TMDBAPI.shared.updateMedia(media, context: managedObjectContext)
-                        await PersistenceController.saveContext(managedObjectContext)
-                    } catch {
-                        Logger.api.error("Failed to update media object: \(error)")
-                    }
-                }
-                dismissEditing()
-            }
-        } label: {
-            Label(Strings.Library.mediaActionReload, systemImage: "arrow.clockwise")
-        }
-    }
-    
-    var deleteButton: some View {
-        Button(role: .destructive) {
-            isShowingDeleteAlert = true
-            AlertHandler.showDeleteAlert(
-                message: Strings.Library.multiDeleteAlertMessage(count: selectedMediaObjects.count)
-            ) {
-                withAnimation {
-                    for media in selectedMediaObjects {
-                        managedObjectContext.delete(media)
-                    }
-                }
-                dismissEditing()
-            }
-        } label: {
-            Label(Strings.Generic.alertDeleteButtonTitle, systemImage: "trash")
-        }
-    }
-    
-    private func dismissEditing() {
-        DispatchQueue.main.async {
-            withAnimation {
-                editMode?.wrappedValue = .inactive
-            }
         }
     }
 }
